Replaced worksheet names with hard-coded values instead of being derived from board name
The board name was being used in the worksheet name, but due to Excel limits (and the use of the 'Archived ' prefix on one tab), that limited the board name length to only 22 characters.  Since the file name is the board name already, the board name doesn't need to be duplicated in the tabs - just have a 'Current' tab and an 'Archived' one.